laba6/requirements.txt
PySimpleGUI
python-docx
openpyxl